[FEAT] #144 온보딩 세팅 완료
