+++ .idea/misc.xml
@@ -25,12 +25,19 @@
         <entry key="..\:/Users/82104/MUMENT-AOS/feature/locker/src/main/res/drawable-hdpi/rectangle_fill_white_top_11dp.xml" value="0.1" />
         <entry key="..\:/Users/82104/MUMENT-AOS/feature/locker/src/main/res/layout/fragment_locker.xml" value="0.23802083333333332" />
         <entry key="..\:/Users/82104/MUMENT-AOS/feature/locker/src/main/res/layout/fragment_locker_filter_bottom_sheet.xml" value="0.1" />
+        <entry key="..\:/Users/82104/MUMENT-AOS/feature/login/src/main/res/drawable/retangle_fill_bgwhite_11dp.xml" value="0.131" />
+        <entry key="..\:/Users/82104/MUMENT-AOS/feature/login/src/main/res/drawable/retangle_fill_purple1_11dp.xml" value="0.137" />
+        <entry key="..\:/Users/82104/MUMENT-AOS/feature/login/src/main/res/drawable/selector_rectangle_white_purple_11dp.xml" value="0.137" />
+        <entry key="..\:/Users/82104/MUMENT-AOS/feature/login/src/main/res/drawable/selector_text_gray_white.xml" value="0.131" />
         <entry key="..\:/Users/82104/MUMENT-AOS/feature/login/src/main/res/font/font.xml" value="0.1824074074074074" />
         <entry key="..\:/Users/82104/MUMENT-AOS/feature/login/src/main/res/layout/activity_log_in.xml" value="0.2" />
-        <entry key="..\:/Users/82104/MUMENT-AOS/feature/login/src/main/res/layout/activity_on_boarding.xml" value="0.264" />
+        <entry key="..\:/Users/82104/MUMENT-AOS/feature/login/src/main/res/layout/activity_on_boarding.xml" value="0.176" />
         <entry key="..\:/Users/82104/MUMENT-AOS/feature/login/src/main/res/layout/activity_splash.xml" value="0.2" />
-        <entry key="..\:/Users/82104/MUMENT-AOS/feature/login/src/main/res/layout/fragment_on_boarding_third.xml" value="0.1453125" />
+        <entry key="..\:/Users/82104/MUMENT-AOS/feature/login/src/main/res/layout/fragment_on_boarding_first.xml" value="0.1" />
+        <entry key="..\:/Users/82104/MUMENT-AOS/feature/login/src/main/res/layout/fragment_on_boarding_second.xml" value="0.125" />
+        <entry key="..\:/Users/82104/MUMENT-AOS/feature/login/src/main/res/layout/fragment_on_boarding_third.xml" value="0.2" />
         <entry key="..\:/Users/82104/MUMENT-AOS/feature/mypage/src/main/res/xml/setting_alarm_preference.xml" value="0.25364583333333335" />
+        <entry key="..\:/Users/82104/MUMENT-AOS/feature/record/src/main/res/drawable-hdpi/retangle_fill_purple1_11dp.xml" value="0.137" />
         <entry key="..\:/Users/82104/MUMENT-AOS/feature/record/src/main/res/drawable/selector_rectagle_gray2_purple1_11dp.xml" value="0.196" />
         <entry key="..\:/Users/82104/MUMENT-AOS/feature/record/src/main/res/layout/fragment_record.xml" value="0.25364583333333335" />
         <entry key="app/src/main/res/drawable-mdpi/selector_ic_list.xml" value="0.1" />
