+++ feature/login/src/main/java/com/mument_android/login/onboarding/OnBoardingThirdFragment.kt
@@ -1,0 +1,23 @@
+package com.mument_android.login.onboarding
+
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.mument_android.login.R
+
+
+class OnBoardingThirdFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_on_boarding_third, container, false)
+    }
+
+
+}
