+++ feature/login/src/main/res/drawable/rectangle_fill_purple1_11dp.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="@color/mument_color_purple1"/>
+    <corners android:radius="11dp"/>
+</shape>
