+++ feature/login/src/main/res/drawable/selector_rectangle_white_purple_11dp.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/rectangle_fill_purple1_11dp" android:state_selected="true" />
+    <item android:drawable="@drawable/rectangle_fill_bgwhite_11dp" android:state_selected="false" />
+</selector>
