+++ feature/login/src/main/res/drawable/selector_text_gray_white.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/white" android:state_selected="true" />
+    <item android:color="@color/mument_color_gray2" android:state_selected="false" />
+</selector>
