+++ feature/login/src/main/res/layout/activity_on_boarding.xml
@@ -35,12 +35,7 @@
             android:layout_height="0dp"
             android:layout_marginTop="45dp"
             app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/dots_indicator" />
-
-
-
-
-
+            app:layout_constraintTop_toBottomOf="@+id/dots_indicator">
+        </androidx.viewpager2.widget.ViewPager2>
     </androidx.constraintlayout.widget.ConstraintLayout>
-
 </layout>
