+++ feature/login/src/main/res/values/themes.xml
@@ -51,6 +51,11 @@
         <item name="android:includeFontPadding">false</item>
     </style>
 
+    <style name="TextView.Mument_H4_m16.Style" parent="">
+        <item name="android:textAppearance">@style/TextView.Mument_H4_m16.TextAppearance</item>
+        <item name="android:includeFontPadding">false</item>
+    </style>
+
 
     <style name="TextView.Mument_H1_B25.TextAppearance" parent="">
         <item name="android:fontFamily">@font/notosans_bold</item>
@@ -102,6 +107,12 @@
         <item name="android:fontFamily">@font/notosans_bold</item>
         <item name="android:textSize">12sp</item>
         <item name="android:textColor">@color/mument_color_gray2</item>
+    </style>
+
+    <style name="TextView.Mument_H4_m16.TextAppearance" parent="">
+        <item name="android:fontFamily">@font/notosans_medium</item>
+        <item name="android:textSize">16dp</item>
+        <item name="android:textColor">@color/mument_color_gray1</item>
     </style>
 
     <!-- case : to use Bottom Sheet -->
